requirements.txt
# Core Dependencies for OpenClass Nexus AI
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=1.1.0
llama-cpp-python>=0.2.20
chromadb>=0.4.15
boto3>=1.34.0
awscli>=1.32.0

# Data Processing
pypdf>=3.17.0
unstructured>=0.11.0
pandas>=2.1.0
numpy>=1.24.0

# ML/AI Libraries
sentence-transformers>=2.2.2
torch>=2.1.0
transformers>=4.35.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
packaging>=23.2
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
hypothesis>=6.92.0

# Development
black>=23.9.0
flake8>=6.1.0